fix: moved plugin-file loading to node-specific validator
