+++ packages/validation/src/node.ts
@@ -1,5 +1,5 @@
 import { ResourceParser } from "./common/resourceParser.js";
-import { loadExternalPlugin, MonokleValidator } from "./MonokleValidator.js";
+import { MonokleValidator } from "./MonokleValidator.js";
 import { SimpleCustomValidator } from "./validators/custom/simpleValidator.js";
 import { SchemaLoader } from "./validators/kubernetes-schema/schemaLoader.js";
 import { KubernetesSchemaValidator } from "./validators/kubernetes-schema/validator.js";
@@ -9,7 +9,6 @@ import { ResourceLinksValidator } from "./validators/resource-links/validator.js
 import { YamlValidator } from "./validators/yaml-syntax/validator.js";
 import fs from "fs";
 import * as path from "path";
-
 
 export function createExtensibleNodeMonokleValidator(
   parser: ResourceParser = new ResourceParser(),
@@ -31,11 +30,32 @@ export function createExtensibleNodeMonokleValidator(
           return new KubernetesSchemaValidator(parser, schemaLoader);
         default:
           try {
-            return await loadExternalPlugin(pluginName, parser);
+            const filePath = path.join(process.cwd(), ".monokle-plugins", `${pluginName}-plugin.js`);
+            if (fs.existsSync(filePath)) {
+              const customPlugin = await import(/* @vite-ignore */ filePath );
+              return new SimpleCustomValidator(customPlugin.default, parser);
+            } else {
+              const url = `https://plugins.monokle.com/validation/${pluginName}/latest.js`;
+              const customPlugin = await importWithDataUrl(url);
+              return new SimpleCustomValidator(customPlugin.default, parser);
+            }
           } catch (err) {
             throw new Error(`plugin_not_found: $err`);
           }
       }
     }
   );
+}
+
+async function importWithDataUrl(url: string) {
+  const response = await fetch(url);
+  if (!response.ok) {
+    throw new Error(`Error fetching ${url}: ${response.statusText}`);
+  }
+  const source = await response.text();
+  const buff = Buffer.from(source);
+  const encodedSource = buff.toString("base64");
+  const dataUrl = `data:text/javascript;base64,${encodedSource}`;
+
+  return await import(/* @vite-ignore */ dataUrl);
 }
